Feature/script workers (#9)
Introduce RabbitMQ as a message broker and a worker service which executes jobs.

+++ scriptinator-data/src/main/java/io/chapp/scriptinator/repositories/AbstractEntityRepository.java
@@ -20,5 +20,5 @@ import org.springframework.data.repository.NoRepositoryBean;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 @NoRepositoryBean
-public interface AbstractEntityRepository<E extends AbstractEntity> extends PagingAndSortingRepository<E, Integer> {
+public interface AbstractEntityRepository<E extends AbstractEntity> extends PagingAndSortingRepository<E, Long> {
 }
